chore(explore): Move explore charts into spans folder
In preparation for a unified chart between spans and logs, move the existing
chart component into the spans folder.
