+++ static/app/views/explore/spans/charts/confidenceFooter.spec.tsx
