+++ static/app/views/explore/spans/charts/confidenceFooter.tsx
