+++ static/app/views/explore/spans/charts/index.spec.tsx
